+++ resources/ReleaseNotes.md
@@ -10,6 +10,7 @@
 - Add operations to plot composite variables (e.g., difference between variables, convert variable value, etc.)
 - Add optional global legend to figures generated with `mag.graphics.visualize`
 - Add option to change tile indexing in figures generated with `mag.graphics.visualize`
+- Add sensor name in "ramp inconsistent" warning
 - Separate `ScienceProcessing` and `IALiRTProcessing` in `mag.IMAPTestingAnalysis`
 - Allow filtering both by time and number of vectors in `mag.process.Filter`
 - Also crop I-ALiRT data when calling `cropScience`